renamed loginAPI to authapi
